Clumsily convert from String to Text
First time directly interacting with the `text` library in more than a
tiny conversion here or there, so I suspect some of the work here could
be done a bit more cleanly.  I'll revisit more going forward, but it
seemed worthwhile to dive into `text` and start getting used to it.