Remove Vector from datatype
I am not sure I am going to bother with fleshing it out as an indexable
datatype, at least not for version 1.  In that case it is easier to drop
it instead of duplicating the list logic for no gain.

Also, updated the README with a rough sketch of my current plans.  That
list might be V1.